Clase 23

+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/dto/salida/PacienteSalidaDto.java
@@ -3,7 +3,7 @@ package com.backend.clinicaodontologica.dto.salida;
 import java.time.LocalDate;
 
 public class PacienteSalidaDto {
-    private int id;
+    private Long id;
     private String nombre;
     private String apellido;
     private int dni;
@@ -13,7 +13,7 @@ public class PacienteSalidaDto {
     public PacienteSalidaDto() {
     }
 
-    public PacienteSalidaDto(int id, String nombre, String apellido, int dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilioSalidaDto) {
+    public PacienteSalidaDto(Long id, String nombre, String apellido, int dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilioSalidaDto) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -22,11 +22,11 @@ public class PacienteSalidaDto {
         this.domicilioSalidaDto = domicilioSalidaDto;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
