+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/service/IPacienteService.java
@@ -12,5 +12,6 @@ public interface IPacienteService {
 
     List<PacienteSalidaDto> listarPacientes();
 
-    PacienteSalidaDto buscarPacientePorId(int id);
+
+    PacienteSalidaDto buscarPacientePorId(Long id);
 }
